task-user

+++ src/tasks/tasks.service.ts
@@ -15,12 +15,12 @@ export class TasksService {
         private taskRepository:TaskRepository
     ){}
 
-    async getTasks(filterDto:GetTaskFilterDto):Promise<Task[]>{
-        return this.taskRepository.getTasks(filterDto)
+    async getTasks(filterDto:GetTaskFilterDto,user:User):Promise<Task[]>{
+        return this.taskRepository.getTasks(filterDto,user)
     }
 
-    async getTaskById(id:number):Promise<Task>{
-        const found = await this.taskRepository.findOne(id)
+    async getTaskById(id:number,user:User):Promise<Task>{
+        const found = await this.taskRepository.findOne({where:{id,userId:user.id}})
         if(!found) throw new NotFoundException(`Task with ID "${id}" not found`)
         return found
     }
@@ -29,15 +29,15 @@ export class TasksService {
         return this.taskRepository.createTask(createTaskDto,user)
     }
     
-    async deleteTask(id:number):Promise<void>{
-        const result =await this.taskRepository.delete(id)
+    async deleteTask(id:number,user:User):Promise<void>{
+        const result =await this.taskRepository.delete({id,userId:user.id})
         if(result.affected === 0) throw new NotFoundException(`Task with ID "${id}" not found`)
-        // console.log('>>>',result);
-        // >>> DeleteResult { raw: [], affected: 1 }
+        // console.log(result);
+        // DeleteResult { raw: [], affected: 1 }
     }
 
-    async updateTaskStatus(id:number,status:TaskStatus):Promise<Task>{
-        const task = await this.getTaskById(id)
+    async updateTaskStatus(id:number,status:TaskStatus,user:User):Promise<Task>{
+        const task = await this.getTaskById(id,user)
         task.status = status;
         await task.save()
         return task
